requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
yfinance>=0.2.28
alpha-vantage>=2.3.1
lxml>=4.9.0
python-dotenv>=1.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
tweepy>=4.14.0
feedparser>=6.0.10

